requirements.txt
feedparser
PyYAML
supabase
requests
python-dateutil
trafilatura>=1.6.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0 